Cast to long to avoid overflow when multiplying by the health (part 2)

+++ OpenRA.Mods.Common/Traits/Buildings/RepairableBuilding.cs
@@ -115,7 +115,9 @@ namespace OpenRA.Mods.Common.Traits
 
 				// The cost is the same regardless of the amount of people repairing
 				var hpToRepair = Math.Min(Info.RepairStep, health.MaxHP - health.HP);
-				var cost = Math.Max(1, (hpToRepair * Info.RepairPercent * buildingValue) / (health.MaxHP * 100));
+
+				// Cast to long to avoid overflow when multiplying by the health
+				var cost = Math.Max(1, (int)(((long)hpToRepair * Info.RepairPercent * buildingValue) / (health.MaxHP * 100L)));
 
 				// TakeCash will return false if the player can't pay, and will stop him from contributing this Tick
 				var activePlayers = Repairers.Count(player => player.PlayerActor.Trait<PlayerResources>().TakeCash(cost, true));
